refactor(game): 클릭 배칭 로직을 useBatchedClick 훅으로 분리
GamePage에 인라인으로 있던 클릭 배칭(5회 또는 500ms 디바운스로
모아 전송) 로직을 별도 훅으로 추출해 페이지 컴포넌트 가독성 확보
